refactoring packages. Tests confirmed work still
