+++ src/main/java/edu/ithaca/dragon/blackjack/Hand.java
@@ -1,4 +1,4 @@
-package com.example.springboot;
+package edu.ithaca.dragon.blackjack;
 
 import java.util.ArrayList;
 import java.util.Collections;
