+++ src/main/java/edu/ithaca/dragon/blackjack/Player.java
@@ -1,7 +1,6 @@
-package com.example.springboot;
+package edu.ithaca.dragon.blackjack;
 
 import java.util.ArrayList;
-import java.util.concurrent.atomic.AtomicInteger;
 
 public class Player implements Comparable {
     private final int ID;
