+++ src/main/java/edu/ithaca/dragon/spring/Application.java
@@ -1,12 +1,8 @@
-package com.example.springboot;
+package edu.ithaca.dragon.spring;
 
-import java.util.Arrays;
-
-import org.springframework.boot.CommandLineRunner;
+import edu.ithaca.dragon.blackjack.Card;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
