+++ src/test/java/edu/ithaca/dragon/blackjack/BlackJackTest.java
@@ -1,11 +1,13 @@
-package com.example.classes;
+package edu.ithaca.dragon.blackjack;
 
-import com.example.springboot.*;
+import edu.ithaca.dragon.blackjack.BlackJack;
+import edu.ithaca.dragon.blackjack.Card;
+import edu.ithaca.dragon.blackjack.Hand;
+import edu.ithaca.dragon.blackjack.Player;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.jupiter.api.Assertions.*;
