+++ src/test/java/edu/ithaca/dragon/blackjack/PlayerTest.java
@@ -1,6 +1,8 @@
-package com.example.classes;
+package edu.ithaca.dragon.blackjack;
 
-import com.example.springboot.*;
+import edu.ithaca.dragon.blackjack.Card;
+import edu.ithaca.dragon.blackjack.Hand;
+import edu.ithaca.dragon.blackjack.Player;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
