+++ src/test/java/edu/ithaca/dragon/spring/HelloControllerTest.java
@@ -1,4 +1,4 @@
-package com.example.springboot;
+package edu.ithaca.dragon.spring;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
